chore: extract main test suite to the world file
